Remove prompt/random exclusivity checks (#96)
in discussion with Michal we agreed that the prompt/random checks should not be exclusive, otherwise it messes up the weighting of the subtraction. If the multiplicity is large and the prompt are random time windows are large, then we would start throwing away events disproportionately. In particular, if the prompt and random windows are different sizes, we get really messed up... Can be reviewed with some data